Data_Retrieval/requirements.txt
boto3
dynamodb_json
flask
matplotlib
numpy
pandas
psycopg2-binary
sqlalchemy
sqlalchemy-redshift
stravalib
